Move Move to combat_AI module.
