+++ src/models/character_base.py
@@ -3,6 +3,8 @@ import logging
 from typing import Sequence
 
 from typing import List
+
+from models.combat_AI import Move
 from models.inventory import BasicInventory, InventoryBase
 from models.states import Attribute, AttributeType, Stateful, BasicStatus, State
 
@@ -55,15 +57,3 @@ class Character(Stateful):
         return moves
 
 
-class Move(object):
-    def __init__(self, ability: 'Ability', user: Character,
-                 target: Character) -> None:
-        self.ability = ability
-        self.user = user
-        self.target = target
-
-    def use(self) -> None:
-        self.ability.use(self.user, self.target)
-
-    def describe(self) -> str:
-        return self.ability.description()
